wip: typing
