+++ src/breds/__init__.py
@@ -1,0 +1,1 @@
+from config import Config
